Organized sql queries into proper files. Modified app to reflect changes. Add flash to appear on registration.
