+++ templates/report_gifts.html
@@ -89,7 +89,7 @@
     <tbody>
         {% if data %}
         <tr>
-            <h3>This table shows how many of an item were sold in a day.</h3>
+            <h3>This table shows how many of an item were sold in a given interval.</h3>
             <th>SKU</th>
             <th>Gift Name</th>
             <th>Price</th>
